feat: configuration studio exports
- exported types and classes required for integration with configuration studio
- removed code duplication in ConfigurationItemManager
