+++ shesha-core/src/Shesha.Framework/ConfigurationItems/ConfigurationItemManager.cs
@@ -205,7 +205,7 @@ namespace Shesha.ConfigurationItems
             return await GetAsync(actualItem.ItemId);
         }
 
-        public virtual async Task<TItem> CreateItemAsync(CreateItemInput input) 
+        public virtual async Task<TItem> CreateItemAsync(CreateItemInput input, object? additionalData = null) 
         {
             var validationResults = new ValidationResults();
             var alreadyExist = await Repository.GetAll().Where(f => f.Module == input.Module && f.Name == input.Name).AnyAsync();
@@ -224,6 +224,8 @@ namespace Shesha.ConfigurationItems
             item.Label = input.Label;
             item.Normalize();
 
+            await BeforeCreateAsync(item, additionalData);
+
             await Repository.InsertAsync(item);
 
             await UnitOfWorkManager.Current.SaveChangesAsync();
@@ -232,48 +234,12 @@ namespace Shesha.ConfigurationItems
         }
 
         /// <summary>
-        /// Creates a configuration item with additional data that can be used during creation process.
-        /// This allows derived classes to pass additional data to HandleAdditionalPropertiesAsync without using temporary fields.
-        /// </summary>
-        /// <param name="input">Basic configuration item properties</param>
-        /// <param name="additionalData">Additional data required by derived classes</param>
-        /// <returns>Created configuration item</returns>
-        public virtual async Task<TItem> CreateItemAsync(CreateItemInput input, object additionalData) 
-        {
-            var validationResults = new ValidationResults();
-            var alreadyExist = await Repository.GetAll().Where(f => f.Module == input.Module && f.Name == input.Name).AnyAsync();
-            if (alreadyExist)
-                validationResults.Add($"Form with name `{input.Name}` already exists in module `{input.Module.Name}`");
-            validationResults.ThrowValidationExceptionIfAny(L);
-
-            var item = new TItem
-            {
-                Name = input.Name,
-                Module = input.Module,
-                Folder = input.Folder,
-            };
-            item.Origin = item;
-            item.Description = input.Description;
-            item.Label = input.Label;
-            item.Normalize();
-
-            // Allow derived classes to handle additional properties with the additional data
-            await HandleAdditionalPropertiesAsync(item, additionalData);
-
-            await Repository.InsertAsync(item);
-            
-            await UnitOfWorkManager.Current.SaveChangesAsync();
-
-            return item;
-        }
-        
-        /// <summary>
-        /// Override this method in derived classes to handle additional properties from CreateItemInput with additional data
+        /// Prepare newly created item
         /// </summary>
         /// <param name="item">The newly created item</param>
         /// <param name="additionalData">Additional data required for specialized item creation</param>
         /// <returns></returns>
-        protected virtual Task HandleAdditionalPropertiesAsync(TItem item, object additionalData)
+        protected virtual Task BeforeCreateAsync(TItem item, object? additionalData)
         {
             // By default, call the simpler overload
             return Task.CompletedTask;
