+++ shesha-reactjs/src/configuration-studio/cs-environment/interfaces.ts
@@ -1,0 +1,17 @@
+import { DocumentDefinition } from "../models";
+
+export interface IConfigurationStudioEnvironment {
+  /**
+   * Register a document definition.
+   */
+  registerDocumentDefinition: (definition: DocumentDefinition) => void;
+  /**
+   * Unregister docuement definition
+   */
+  unregisterDocumentDefinition: (definition: DocumentDefinition) => void;
+
+  /**
+   * Get document definition by item type
+   */
+  getDocumentDefinition: (itemType: string) => DocumentDefinition | undefined;
+}
