+++ shesha-reactjs/src/configuration-studio/cs/contexts.tsx
@@ -1,5 +1,6 @@
 import { createNamedContext } from "@/utils/react";
-import { ConfigurationStudio, IConfigurationStudio } from "./configurationStudio";
+import { ConfigurationStudio } from "./configurationStudio";
+import { IConfigurationStudio } from "./interfaces";
 import React, { FC, PropsWithChildren, useContext, useEffect, useLayoutEffect, useRef } from "react";
 import { useHttpClient, useShaRouting } from "@/providers";
 import { asyncStorage } from "../storage";
@@ -7,12 +8,14 @@ import { useModalApi } from "./modalApi";
 import { useNotificationApi } from "./notificationApi";
 import { isDefined } from "../../utils/nullables";
 import { useSearchParams } from "next/navigation";
+import { useConfigurationStudioEnvironment } from "../cs-environment/contexts";
 
 export const useIsomorphicLayoutEffect =
   typeof window !== "undefined" ? useLayoutEffect : useEffect;
 
-const useConfigurationStudioSingletone = (): IConfigurationStudio[] => {
+const useConfigurationStudioSingletone = (): [IConfigurationStudio] => {
   const [, forceUpdate] = React.useState({});
+  const csEnvironment = useConfigurationStudioEnvironment();
   const httpClient = useHttpClient();
   const modalApi = useModalApi();
   const notificationApi = useNotificationApi();
@@ -24,6 +27,7 @@ const useConfigurationStudioSingletone = (): IConfigurationStudio[] => {
     // Create a new FormStore if not provided
     const instance = new ConfigurationStudio({
       forceRootUpdate: () => forceUpdate({}),
+      csEnvironment,
       httpClient,
       storage: asyncStorage,
       modalApi: modalApi,
