+++ shesha-reactjs/src/configuration-studio/cs/hooks.ts
@@ -1,6 +1,6 @@
 import { ReactNode, useEffect, useState } from "react";
-import { CloseDocumentResponse, ConfigItemTreeNode, DocumentDefinition, IDocumentInstance, TreeNode } from "../models";
-import { CsSubscriptionType, ProcessingState } from "./configurationStudio";
+import { CloseDocumentResponse, ConfigItemTreeNode, IDocumentInstance, TreeNode } from "../models";
+import { CsSubscriptionType, ProcessingState } from "./interfaces";
 import { useConfigurationStudio, useConfigurationStudioIfAvailable } from "./contexts";
 import { TreeProps } from "antd";
 import { isDefined } from "../../utils/nullables";
@@ -113,18 +113,4 @@ export const useActiveDoc = (): UseActiveDocResponse => {
   return cs.activeDocument;
 };
 
-export const useConfigurationStudioDocumentDefinitions = (definitions: DocumentDefinition[]): void => {
-  const cs = useConfigurationStudio();
 
-  useEffect(() => {
-    definitions.forEach((definition) => {
-      cs.registerDocumentDefinition(definition);
-    });
-
-    return (): void => {
-      definitions.forEach((definition) => {
-        cs.unregisterDocumentDefinition(definition);
-      });
-    };
-  }, [cs, definitions]);
-};
