+++ shesha-reactjs/src/configuration-studio/menu-utils.tsx
@@ -3,7 +3,7 @@ import { MenuProps } from "antd";
 import React from "react";
 import { ConfigItemTreeNode, FolderTreeNode, isConfigItemTreeNode, isFolderTreeNode, isModuleTreeNode, ModuleTreeNode, TreeNode, TreeNodeType } from "./models";
 import { getIcon } from "./tree-utils";
-import { IConfigurationStudio } from "./cs/configurationStudio";
+import { IConfigurationStudio } from "./cs/interfaces";
 import { isDefined } from "../utils/nullables";
 
 type MenuItemType = NonNullable<MenuProps["items"]>[number];
