+++ shesha-reactjs/src/index.tsx
@@ -33,3 +33,8 @@ export { useResultTypeEvaluator } from './designer-components/codeEditor/hooks/u
 export { default as ConfigurationStudio } from './configuration-studio';
 export { useFormViaFactory, useFormBuilderFactory } from './form-factory/hooks';
 export { type FormBuilder, type FormBuilderFactory } from './form-factory/interfaces';
+
+export { type DocumentDefinition, type IDocumentInstance } from './configuration-studio/models';
+export { DocumentInstance } from './configuration-studio/cs/documentInstance';
+export { DocumentDefinitionRegistration } from './configuration-studio/document-definitions/documentDefinitionRegistration';
+export { useConfigurationStudio } from './configuration-studio/cs/contexts';
